moving necessary files from the `presentation` module to the `features:jumps:presentation` module
