+++ features/jumps/presentation/src/main/java/com/fredprojects/features/jumps/presentation/JDCustomViews.kt
@@ -1,4 +1,4 @@
-package com.fredprojects.helloworld.presentation.features.jumps
+package com.fredprojects.features.jumps.presentation
 
 import androidx.compose.foundation.border
 import androidx.compose.foundation.layout.*
@@ -9,10 +9,9 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.unit.dp
-import com.fredprojects.helloworld.presentation.R
-import com.fredprojects.helloworld.domain.core.utils.SortType
-import com.fredprojects.helloworld.presentation.core.Action
-import com.fredprojects.helloworld.presentation.core.FredRadioButton
+import com.fredprojects.core.ui.*
+import com.fredprojects.core.ui.R
+import com.fredprojects.features.jumps.domain.utils.SortType
 
 @Composable
 internal fun JDSortingSection(
@@ -28,7 +27,11 @@ internal fun JDSortingSection(
 internal fun JDIconButton(onClick: Action, icon: ImageVector) {
     IconButton(
         onClick,
-        Modifier.padding(4.dp).border(2.dp, MaterialTheme.colors.onBackground, MaterialTheme.shapes.small)
+        Modifier.padding(4.dp).border(
+            width = 2.dp,
+            color = MaterialTheme.colors.onBackground,
+            shape = MaterialTheme.shapes.small
+        )
     ) {
         Icon(icon, icon.toString())
     }
