+++ features/jumps/presentation/src/main/java/com/fredprojects/features/jumps/presentation/StickMan.kt
@@ -1,4 +1,4 @@
-package com.fredprojects.helloworld.presentation.features.jumps
+package com.fredprojects.features.jumps.presentation
 
 import androidx.compose.animation.core.*
 import androidx.compose.foundation.Canvas
