+++ features/jumps/presentation/src/main/java/com/fredprojects/features/jumps/presentation/vm/JDEvents.kt
@@ -1,7 +1,7 @@
-package com.fredprojects.helloworld.presentation.features.jumps.vm
+package com.fredprojects.features.jumps.presentation.vm
 
-import com.fredprojects.helloworld.domain.core.utils.SortType
-import com.fredprojects.helloworld.domain.features.jumps.models.JumpData
+import com.fredprojects.features.jumps.domain.models.JumpData
+import com.fredprojects.features.jumps.domain.utils.SortType
 
 /**
  * JDEvents is used to handle events in the JumpDataList
@@ -15,6 +15,6 @@ sealed class JDEvents {
     data class Sort(val sortType: SortType) : JDEvents()
     data class DeleteJD(val jumpData: JumpData) : JDEvents()
     data class UpsertJD(val jumpData: JumpData) : JDEvents()
-    data class GetJD(val jumpDataId: Int) : JDEvents()
+    data class GetJD(val jumpData: JumpData) : JDEvents()
     data object ToggleSortSection : JDEvents()
 }
